Add files via upload
The objective of this project is to practice what have been learned in the section of data wrangling and anaysis from “Udacity Data Analysis Nanodegree Program”. Through this
project, the dataset wrangled, analyzed, and visualized is the tweet archive of Twitter user @dog_rates which is also known as WeRateDogs. Briefly, WeRateDogs is a Twitter
account that rates people's dogs with humorous comments about the dog. Further information about this issue can be seen at Wikipedia or at Twitter